fixed storage (added new field ID - PK)

+++ src/main/java/ru/naburnm8/bmstu/datamanagementnirbackend/controllers/CatalogueController.java
@@ -5,8 +5,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import ru.naburnm8.bmstu.datamanagementnirbackend.models.Catalogue;
+import ru.naburnm8.bmstu.datamanagementnirbackend.models.Storage;
 import ru.naburnm8.bmstu.datamanagementnirbackend.services.CatalogueService;
+import ru.naburnm8.bmstu.datamanagementnirbackend.services.StorageService;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -16,6 +19,9 @@ public class CatalogueController {
 
     @Autowired
     private CatalogueService catalogueService;
+
+    @Autowired
+    private StorageService storageService;
 
     @GetMapping
     public List<Catalogue> getAllCatalogues() {
@@ -30,7 +36,12 @@ public class CatalogueController {
 
     @PostMapping
     public Catalogue createCatalogue(@RequestBody Catalogue catalogue) {
-        return catalogueService.createCatalogue(catalogue);
+        Catalogue added = catalogueService.createCatalogue(catalogue);
+        ArrayList<Catalogue> catalogues = new ArrayList<>(catalogueService.getAllCatalogues());
+        for(Catalogue c: catalogues) {
+            storageService.createIfNotPresent(c);
+        }
+        return added;
     }
 
     @PutMapping("/{id}")
@@ -40,7 +51,11 @@ public class CatalogueController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteCatalogue(@PathVariable int id) {
-        catalogueService.deleteCatalogueById(id);
+        Optional<Catalogue> catalogue = catalogueService.getCatalogueById(id);
+        if (catalogue.isPresent()) {
+            storageService.deleteByCatalogue(catalogue.get());
+            catalogueService.deleteCatalogueById(id);
+        }
         return ResponseEntity.ok().build();
     }
 }
